+++ eggs/Naaya/CHANGELOG.rst
@@ -1,4 +1,4 @@
-3.3.20 (unreleased)
+3.3.20 (2013-11-29)
 -------------------
 * Updated naaya.core.ggeocoding to use GoogleMaps api v3
   [tiberich #16938]